Store data in .data, implement localChange event
Previously, the data had to be computed from the whole history upon access. We
now update it internally on every operation.

Also, we plan on using a parameter, disableData, for use when clients don't need
to hold on to the data. Typically, this is the case when all user input is
handled by widgets (instead of widgets being updated when the localChange event
fires, after user input was handled by canop). When the client does not hold
data, it only needs sent and local histories, for rebasing purposes. The data is
then in the widgets, but not queriable.