Automatically remove server-side operations when acked by all clients
It required a tweak to the protocol, so that we now have three defined protocol
states for clients: unsynced, loading (either a snapshot or a history), ready.
Indeed, now receiving a delta does not imply that the client is ready, as that
delta can be sent from a ready node, while this node may not be ready.

Also, reset the client-side reconnection interval when the connection is back
up.