Support multi-cursors with beforeChange
Commit 49a2c1eb54687dfb4332a37f4912e0fe1e006533 attempted to support
multi-cursors by relying on the 'changes' event, which yields multiple
operations, but after they have been applied to the editor.

However, the CodeMirror positions of these operations correspond to
before the change's execution, which made for a convoluted process to
rebase indices properly.

Another option was to rely on the 'beforeChange' event, which explicitly
triggers before the changes are applied to CodeMirror. It hugely
simplifies the code.

I verified that it yields correct results with the following sequence of
operations, which failed with the old way of relying on the 'change'
event:

1. Write "ab\ncd".
2. Put two cursors, each at the start of each line, and select the first character of the line.
3. Cut the selection.
4. Paste the selection at the end of the line (still with two cursors).